FIX Removing applied filters

+++ GraphVisualizer/graph_explorer/webapp/views.py
@@ -73,34 +73,39 @@ def simple_visualizer(request):
     context["filter_string"] = _applied_filters()
     return render(request, "simple_template.html", context)
 
-def load_file(request):
-     if request.method == "POST":
-        uploaded_file = request.FILES.get("load_file")
-        if uploaded_file:
-            uri = "neo4j://127.0.0.1:7687"
-            username = "neo4j"
-            password = "djomlaboss"
+def load_file(request=None):
+    if request is not None:
+        if request.method == "POST":
+            uploaded_file = request.FILES.get("load_file")
+            if uploaded_file:
+                uri = "neo4j://127.0.0.1:7687"
+                username = "neo4j"
+                password = "djomlaboss"
 
-            # Kreiranje drajvera
-            driver = GraphDatabase.driver(uri, auth=(username, password))
+                # Kreiranje drajvera
+                driver = GraphDatabase.driver(uri, auth=(username, password))
 
-            file_type = uploaded_file.name.split(".")[1]
-            # Napravi privremeni fajl sa suffix .yaml
-            with tempfile.NamedTemporaryFile(delete=False, suffix=file_type) as tmp:
-                for chunk in uploaded_file.chunks():
-                    tmp.write(chunk)
-                tmp_path = tmp.name  # privremena putanja fajla
+                file_type = uploaded_file.name.split(".")[1]
+                # Napravi privremeni fajl sa suffix .yaml
+                with tempfile.NamedTemporaryFile(delete=False, suffix=file_type) as tmp:
+                    for chunk in uploaded_file.chunks():
+                        tmp.write(chunk)
+                    tmp_path = tmp.name  # privremena putanja fajla
 
-            parser = YamlFileParser(tmp_path, driver)
-            parser.load()
-            # Zatvaranje drajvera
-            driver.close()
+                parser = YamlFileParser(tmp_path, driver)
+                parser.load()
+                # Zatvaranje drajvera
+                driver.close()
 
-            # Možeš odmah obrisati privremeni fajl
-            os.remove(tmp_path)
-            return simple_visualizer(request)
+                # Možeš odmah obrisati privremeni fajl
+                os.remove(tmp_path)
+                
+    if current_view == "simple":
+        return simple_visualizer(request)
+    else:
+        return block_view(request)
 
-def make_search(request):
+def make_search(request=None):
     if request.method == "POST":
         search_query = request.POST.get("search", "").strip()
 
@@ -123,32 +128,51 @@ def make_search(request):
             return block_view(request)
 
 
-def apply_filter(request):
-     if request.method == "POST":
-        filter_attribute = request.POST.get("filter", "").strip()
-        filter_relation = request.POST.get("relations", "").strip()
-        filter_value = request.POST.get("value", "").strip()
+def apply_filter(request=None):
+    if request is not None:
+        if request.method == "POST":
+            filter_attribute = request.POST.get("filter", "").strip()
+            filter_relation = request.POST.get("relations", "").strip()
+            filter_value = request.POST.get("value", "").strip()
 
-        if filter_attribute != "" and filter_relation != "" and filter_value != "":
-            try:
-                actual_value = float(filter_value)
-            except:
-                actual_value = filter_value
+            if filter_attribute != "" and filter_relation != "" and filter_value != "":
+                try:
+                    actual_value = float(filter_value)
+                except:
+                    actual_value = filter_value
+                
+                single_filter = (filter_attribute, filter_relation, actual_value)
+                if single_filter not in filters:
+                    filters.append(single_filter)
             
-            single_filter = (filter_attribute, filter_relation, actual_value)
-            if single_filter not in filters:
-                filters.append(single_filter)
-        
-        if current_view == "simple":
-            return simple_visualizer(request)
-        else:
-            return block_view(request)
+    if current_view == "simple":
+        return simple_visualizer(request)
+    else:
+        return block_view(request)
     
 def _applied_filters():
     list_strings = []
+    list_strings_to_file = []
     for filter in filters:
         list_strings.append("".join(str(x) for x in filter))
+        list_strings_to_file.append("~".join(str(x) for x in filter))
     
     filter_string = ",".join(list_strings)
-    return filter_string
+    filter_string_to_file = ",".join(list_strings_to_file)
+    with open("session.txt", "w+") as f:
+        f.write(filter_string_to_file + "\n")
+        f.write(current_view)
 
+    return list_strings
+
+def filter_remove(request):
+    selected_filter = request.GET.get("filter")
+    for single_filter in filters:
+        merged = "".join(str(x) for x in single_filter)
+        if selected_filter == merged:
+            filters.remove(single_filter)
+            break
+    if current_view == "simple":
+        return simple_visualizer(request)
+    else:
+        return block_view(request)
